pipelines: move pipeline utils to pipelines/utils
